requirements.txt
openenv-core
pandas
numpy
scikit-learn
tabulate
fireducks
uvicorn
fastapi
pydantic
